Convert whole document to ASCII
MongoDB returns documents via JSON which properly decodes
to unicode.  However Sybase requires us to use Python2 which
doesn't use unicode for its strings.  This commit replaces the
piecemeal conversion of FITS headers to ASCII as and when required
with an encoding of all the values in the document returned
by MongoDB.  This should make the ingestion code simpler
and less error prone.