Add support for control files.
These files record the list of observations which have already
been ingested, and therefore allow the ingestion process
to be re-run multiple times to complete the ingestion.
They work in the same way as those used in the ukirt-archive
scripts for transferring the files to the archive by
eTransfer, i.e. when an observation is ingested, its name
is written into the control file.  Then this file is read
into a Python set for the next ingestion run, and observations
are skipped if they are already in the set.