Add new archive job submission script
This is based on and replaces the old "prepare_dr.py" script.  The
useful parts of the DR_Prep class have been placed in two new
classes, ObsList and HeaderOverride.  Only the first of these is
used so far by the submission script as the current focus is on
UIST which should not require header overrides.