Update filter names
Attempted to tidy up the filter names.  I had previously found
"official" filter names from the UKIRT instrument webpages and
translated the names from the FITS headers to those names.
However it seems that the filters were inconsistently named
on the webpages and so the filter pick list on CADC's Advanced
Search interface is quite muddled.  Watson indicated that the
"98" and "MK" sets are the same thing.

Also added filter names for CGS3 and CGS4.